- moved addCosignatures method from AggregateTransaction into AggregateTransactionService to get rid of circular dependency #66 - more missing exports added
